Huge commit which fixes and breaks stuff
Re-align stuff in dialogs. Make app remember last used status (not text,
but this online/offline thingy) , so it can be reused while
reconnecting, or just changing status dialog. Also, stub-plement
Symbiosis Notification API. I want Lightbulb to be the first app to
support it and I'm gonna test it here. So it becomes a freakin'
minefield now :P