Improve archive view
Because of the bug in ListView, I can't make it scroll to the end if
there are too many messages (view gets positioned somewhere in the
middle and there is no way of changing that) so I got them all sliced
into pages of 20 messages. WIP implementation, can't test right now.