Fix QDnsLookup on Symbian
I just commented out some code and linked it against some libs so it's
actually building. Not sure if it works but this code looks kinda
smart... nevermind.

I also added some stuff in MyXmppClient to use SSL and TLS. About the UI
- donation button is here for anyone who has too much money and wants to
show some support for my work and I decided it would be awesome to
mention Pauls contribution.