Support unread count, sort of
First, this code won't really check if contact is on the list and won't
add it if necessary. And as it doesn't really check if account is
correct, this implementation is very name-clashing prone. I don't wanna
know what happens if you have a contact with the same JID on multiple
accounts.