Trying to clean up the VCard mess and failed
So, the good thing is, I made contact menu care about accountId. This
can be later pushed to 0.4 where it would be much more useful. I thought
about using the code from 0.4 but - sure, it was done after a sleepless
night and during a 9 hour roadtrip so - no, it's shit. I tried to touch
the VCard code but I guess I'm to sober to get through this mess. Well,
at least it works, right? I need to rewrite it, maybe use SQLite backend
here. The interesting thing is, app saved VCards but it never read the
crap that was saved. This part of code was derived from MeegIM and never
got that much love as other parts so I never knew how it works. I still
don't, to be honest. These changes could be merged to 0.4 as well, where
they'll be much more useful than here.