New haptics effect code
It plays like four times if you receive a message, vibrations don't
overlap themselves, it makes a short beep when the app is active. If you
get multiple messages while notification is playing, it still will play
just 4 times, not 4 + 4 + 4 for every message. If you switch to the app
while the long vibration sound is playing, it will stop.